[Decomp][Spell] Cohesion split: extract effect dispatch to SpellEffectDispatch.cpp
